+++ src/preload.js
@@ -13,6 +13,7 @@ contextBridge.exposeInMainWorld('waveframe', {
   exportEnd: (id) => ipcRenderer.invoke('export-end', id),
   exportCancel: (id) => ipcRenderer.invoke('export-cancel', id),
   showInFolder: (filePath) => ipcRenderer.invoke('show-in-folder', filePath),
+  appInfo: () => ipcRenderer.invoke('app-info'),
   updateDownload: () => ipcRenderer.invoke('update-download'),
   updateInstall: () => ipcRenderer.invoke('update-install'),
   onUpdateState: (callback) => {